extract QueryFactoryFunctionalTest
