+++ src/test/kotlin/r2dbcfun/test/functional/RepositoryFunctionalTest.kt
@@ -1,22 +1,15 @@
 package r2dbcfun.test.functional
 
 import io.kotest.core.spec.style.FunSpec
-import kotlinx.coroutines.flow.single
-import kotlinx.coroutines.flow.toCollection
 import kotlinx.coroutines.reactive.awaitSingle
 import kotlinx.serialization.Serializable
 import r2dbcfun.NotFoundException
 import r2dbcfun.PK
 import r2dbcfun.Repository
 import r2dbcfun.forAllDatabases
-import r2dbcfun.query.between
-import r2dbcfun.query.isEqualTo
-import r2dbcfun.query.isNull
-import r2dbcfun.query.like
 import r2dbcfun.util.toSnakeCase
 import strikt.api.expectCatching
 import strikt.api.expectThat
-import strikt.assertions.containsExactly
 import strikt.assertions.isA
 import strikt.assertions.isEqualTo
 import strikt.assertions.isFailure
@@ -109,78 +102,6 @@ class RepositoryFunctionalTest : FunSpec({
                             get { birthday }.isEqualTo(LocalDate.parse("2020-06-20"))
                         }
                 }
-                context("query language") {
-                    test("has a typesafe query api") {
-                        // create 3 users with different birthdays so that only the
-                        // middle date
-                        // fits the between condition
-                        create(
-                            User(
-                                name = "chris",
-                                email = "my email",
-                                birthday = LocalDate.parse("2020-06-18")
-                            )
-                        )
-                        val userThatWillBeFound =
-                            create(
-                                User(
-                                    name = "jakob",
-                                    email = "different email",
-                                    birthday = LocalDate.parse("2020-06-20")
-                                )
-                            )
-                        create(
-                            User(
-                                name = "chris",
-                                email = "different email",
-                                birthday = LocalDate.parse("2020-06-22")
-                            )
-                        )
-                        val date1 = LocalDate.parse("2020-06-19")
-                        val date2 = LocalDate.parse("2020-06-21")
-                        val findByUserNameLikeAndBirthdayBetween =
-                            repo.queryFactory
-                                .createQuery(User::name.like(), User::birthday.between())
-
-                        expectThat(
-                            findByUserNameLikeAndBirthdayBetween(
-                                connection,
-                                "%",
-                                Pair(date1, date2)
-                            ).toCollection(mutableListOf())
-                        ).containsExactly(userThatWillBeFound)
-                    }
-                    test("can query null values") {
-                        val coolUser =
-                            create(User(name = "coolUser", email = "email", isCool = true))
-                        val uncoolUser =
-                            create(
-                                User(name = "uncoolUser", email = "email", isCool = false)
-                            )
-                        val userOfUndefinedCoolness =
-                            create(
-                                User(
-                                    name = "userOfUndefinedCoolness",
-                                    email = "email",
-                                    isCool = null
-                                )
-                            )
-                        val findByCoolness =
-                            repo.queryFactory.createQuery(User::isCool.isEqualTo())
-                        val findByNullCoolness =
-                            repo.queryFactory.createQuery(User::isCool.isNull())
-                        expectThat(findByNullCoolness(connection, Unit).single())
-                            .isEqualTo(userOfUndefinedCoolness)
-                        // this does not compile because equaling by null makes no sense
-                        // anyway:
-                        // expectThat(findByCoolness(connection,
-                        // null).single()).isEqualTo(userOfUndefinedCoolness)
-                        expectThat(findByCoolness(connection, false).single())
-                            .isEqualTo(uncoolUser)
-                        expectThat(findByCoolness(connection, true).single())
-                            .isEqualTo(coolUser)
-                    }
-                }
 
                 test("throws NotFoundException when id does not exist") {
                     expectCatching { repo.findById(connection, UserPK(1)) }.isFailure()
@@ -271,6 +192,7 @@ class RepositoryFunctionalTest : FunSpec({
                     }
             }
         }
+
     }
 }
 )
